Added comments and formatted code

+++ src/models/SEB.jl
@@ -1,20 +1,31 @@
 
 function M(descs::Dict)
+    #Calculating molecular weight of given molecul
     return get(descs,"exactmw",0)
 end
+
 function R(descs::Dict)
+    #Calculating boolean parameter determing the presence of rings of given molecul
     return !(get(descs,"NumRings",0) == 0)
 end
+
 function r_het(descs::Dict)
+    #Calculating the ratio of Number of Heteroatoms to number of heavy atoms of given molecul
     return (get(descs,"NumHeteroatoms",0)/get(descs,"NumHeavyAtoms",0))
 end
+
 function r_acc(descs::Dict)
+    #Calculating ratio of hydrogenbondacceptors to number of heavy atoms of given molecul
     return (get(descs,"NumHBA",0)/get(descs,"NumHeavyAtoms",0))
 end
+
 function r_don(descs::Dict)
+    ##Calculating ratio of hydrogenbonddonors to number of heavy atoms of given molecul
     return (get(descs,"NumHBD",0)/get(descs,"NumHeavyAtoms",0))
 end
+
 function r_hal(SMILES::String,descs::Dict)
+    #Calculating ratio of halogenatoms to heavy atoms of given molecule
     mol = get_mol(SMILES)
     N_h=0
     for h in ["F", "Cl", "Br", "I"] # , "At", "Ts"]
@@ -56,8 +67,11 @@ struct SEB{M}
 end
 
 function SEB(SMILE_i::String,SMILE_j::String,eta_fun)
+    #Loading weights and bias from nn_parameters.jld2
     path_nn_parameters = joinpath(DB_PATH, "SEB", "nn_parameters.jld2")
     nn_parameters = load(path_nn_parameters)["Weights_Bias_SEB"]
+    
+    #Processing SMILES to get molecular descriptors used in neural net
     mol_i,mol_j = get_mol(SMILE_i), get_mol(SMILE_j)
     desc_i,desc_j = get_descriptors(mol_i), get_descriptors(mol_j)
     X_i_ini=[M(desc_i);R(desc_i);r_het(desc_i);r_hal(SMILE_i,desc_i);r_acc(desc_i);r_don(desc_i)]
@@ -68,21 +82,28 @@ function SEB(SMILE_i::String,SMILE_j::String,eta_fun)
     #b_ij Berechnung => Modell
     b_ij=0
     b_ij_sum=0
+
+    #Initializing Neural Net using Lux
     NN = Chain(Dense(12 => 32, relu),Dense(32 => 16, relu),Dense( 16 => 1, softplus))
     #keys=["SEB_3";"SEB_7";"SEB_9";"SEB_12";"SEB_17";"SEB_19";"SEB_33";"SEB_42";"SEB_49";"SEB_55"]
+
+    # Looping over parameter-sets to determine mean b_ij
     for (key,wb) in nn_parameters
+
+        #Setting up the weights and bias of the neural net using Lux-Synatx
         st_0=(layer_1=NamedTuple(),layer_2=NamedTuple(),layer_3=NamedTuple())
         ps=((layer_1=(weight=wb[2],bias=vec(wb[1]))),
         (layer_2=(weight=wb[4],bias=vec(wb[3]))),
         (layer_3=(weight=wb[6],bias=vec(wb[5]))))
 
+        #applying neural net with given weights and bias to calculate b_ij
         b_ij, st = NN(input,ps,st_0)
         b_ij_sum = b_ij_sum+b_ij[1]
 
     end
     b_ij_mean=b_ij/10
-    #setup()
 
+    # Constructing SEBParam-datastructure
     paramSEB=SEBParam(MW,b_ij_mean)
 
 
@@ -91,7 +112,8 @@ end
 
 Base.broadcastable(x::SEB) = Ref(x)
 
-function Diffusion(model::SEB,p::Float64,T::Float64)
+function Diffusion(model::SEB,p,T)
+    # Initialitzing constants required for Stokes-Einstein-equation
     k_b, roh_i, f, M_i,visc_j,N_A = 1.380649*10^(-23), 1050, 0.64, model.param.M, model.vis_model(T),6.02214076*10^23
 
     D_SEE_ij_infdil = (k_b*T)/(6*pi*visc_j*(((3*f*M_i)/(4*pi*roh_i*N_A))^(1/3)))
